style(subway): separate the class
- 하나의 클래스를 여러 개의 클래스로 분리
