+++ src/main/java/subway/view/output/ExceptionView.java
@@ -1,0 +1,10 @@
+package subway.view.output;
+
+import subway.type.ExceptionType;
+
+public class ExceptionView {
+    public static void printInvalidFeatureChoiceException() {
+        System.out.println();
+        System.out.println(ExceptionType.INVALID_FEATURE_CHOICE.getException());
+    }
+}
