+++ src/main/java/subway/view/output/TextView.java
@@ -1,0 +1,10 @@
+package subway.view.output;
+
+import subway.type.TextType;
+
+public class TextView {
+    public static void printStationAddingText() {
+        System.out.println();
+        System.out.println(TextType.STATION_ADDING_TEXT.getText());
+    }
+}
